requirements.txt
Flask 
python-nmap